requirements.txt
mss
numpy
opencv-python
PyYAML
winotify
